deeptrace_backend/deeptrace-backend/requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
flask-cors==6.0.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.1
pillow==12.1.0
pillow_heif==1.1.1
pycryptodome==3.23.0
Werkzeug==3.1.4
